requirements.txt
Flask
requests
python-dotenv
openai
bootstrap-flask
markdown2
google-api-python-client
google-auth
gunicorn
playwright
